FIX HotzonesDistribute() and add RollDice2
We can not just add 'echo "$DICE"' to RollDice() - it'll break a half
of code. So I think use RollDice2() when we need return from function
and bit by bit replace RollDice() to RollDice2(). When all invocations
of RollDice() will be replaced we will need just delete RollDice() and
rename RollDice2() to RollDice() (and of course fix all invocations).